Change send/recv data to uint8_t

+++ send/main.cpp
@@ -7,7 +7,7 @@ int main()
     
     sock.bind("tcp://127.0.0.1:44100");
 
-    std::vector<int> ivec = { 10, 20, 30 };
+    std::vector<uint8_t> ivec = { 10, 20, 30 };
     zmq::message_t msg(ivec);
     sock.send(msg, zmq::send_flags::none);
 }
